Apply code suggestions Use size_t for port Assert if _interruptInstalledMouse becomes less than zero nullptr initializer for _client Add delay back in for INFO_LOG

+++ VoodooPS2Controller/ApplePS2Device.cpp
@@ -29,7 +29,7 @@ OSDefineMetaClassAndStructors(ApplePS2Device, IOService);
 // ApplePS2Device Class Implementation
 //
 
-bool ApplePS2Device::init(int port)
+bool ApplePS2Device::init(size_t port)
 {
   _port = port;
   return super::init();
